requirements.txt
fastapi
uvicorn
tortoise-orm
asyncpg
aerich
numpy